requirements.txt
streamlit
opencv-python
numpy
ultralytics
Pillow
tensorflow
huggingface_hub
streamlit-webrtc
av
